src/requirements.txt
win32clipboard
socket
threading
requests
pynput.keyboard
os
subprocess
shututil